requirements.txt
discord
asyncio
datetime
mariadb
discord_components
aiohttp
aioping
pillow
pyfiglet
ytdl
voice
discord-py-slash-command
pytz
colorama